Mobile tweaks 2 (#281)
* fix scan results table

* fix scan results table

* constraint template editor now scrolls instead of wrapping

* kubesec report page now looks good mobile

* gatekeeper now actually cares whether you confirm or cancel in delete confirmation dialog

* gatekeeper install dialog copy button now responsive

* vuln history table responsiveness fix

* fix namespace in vuln difference by date

* clean up vuln difference by date table

* cleanup

+++ dash/frontend/src/app/modules/private/pages/cluster/gate-keeper-details/gate-keeper-details.component.ts
@@ -85,16 +85,18 @@ export class GateKeeperDetailsComponent implements OnInit {
 
     openDestroyTemplateDialog.afterClosed()
       .pipe(take(1))
-      .subscribe(response => {
-        this.gatekeeperService.deleteConstraintTemplate(this.clusterId, this.templateName)
-          .pipe(take(1))
-          .subscribe(() => {
-            this.alertService.success('Constraint Template successfully deleted');
-            this.router.navigate([`private/clusters/${this.clusterId}/gatekeeper`]);
-          }, error => {
-            this.alertService.dangerAlertForHTTPError(error, 'GateKeeperDetailsComponent.destroyConstraintTemplate');
-            this.getConstraintTemplateDetails();
-          });
+      .subscribe((confirmed: boolean) => {
+        if (confirmed) {
+          this.gatekeeperService.deleteConstraintTemplate(this.clusterId, this.templateName)
+            .pipe(take(1))
+            .subscribe(() => {
+              this.alertService.success('Constraint Template successfully deleted');
+              this.router.navigate([`private/clusters/${this.clusterId}/gatekeeper`]);
+            }, error => {
+              this.alertService.dangerAlertForHTTPError(error, 'GateKeeperDetailsComponent.destroyConstraintTemplate');
+              this.getConstraintTemplateDetails();
+            });
+        }
       });
   }
